Refs #3567 - Generate rosidl_typesupport_microxrcedds test. Serialized and deserialized test. Id test. Nesting msg test.

+++ rosidl_typesupport_microxrcedds_c/src/deserialize_buffer_utility.c
@@ -20,7 +20,7 @@
 
 void rosidl_typesupport_microxrcedds_c__align_pointer(uint8_t** pointer_to_align, size_t *mem_size)
 {
-	size_t word_size = sizeof(int);
+	size_t word_size = sizeof(uint64_t);
 	uint8_t* aling_pointer = (uint8_t*)((((size_t)(*pointer_to_align)) / word_size) * word_size);
 
 	if (*pointer_to_align > aling_pointer)
